Disable sdt.
We don't list dtrace in DEPENDS so we shouldn't be depending on this header.
It may or may not exist from preivous builds though. To be determinstic, disable
sdt.h usage always. This avoids build failures if the header is removed after configure
but before libgcc is compiled for example.

RP 2012/8/7

Signed-off-by: Khem Raj <raj.khem@gmail.com>

Disable sdt for libstdc++-v3.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>

Upstream-Status: Inappropriate [hack]