gcc/config/microblaze: Make r21 fixed due to Linux ABI
Commit 6dcad60c0ef48af584395a40feeb256fb82986a8 introduced a change to
how the r20 and r21 registers are handled for MicroBlaze. Specifically
removing the fixed nature of the r21 register.

In the Linux ABI the r21 register is reserved as a fixed register for
the purposes of storing a pointer to the thread related object. This is
of particular importance for libc (glibc and musl) which use this
register and expect it to be fixed for this purpose.

GCC uses the r20 register for GOT pointing. This commit mentioned above
confusingly states it preserved the fixed nature but in fact changes r20
to be non-fixed. This patch fixes that up.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>