microblaze/sync.md: Correct behaviour and define side-effects
This change corrects the behaviour with regards to the bool output.
Previously the definition would set the bool operand to true (non-zero)
on failure, specifically at the 'cmp' against the expected operand which
would be set non-zero when the memory != expected value. Instead of
using the bool operand as the compare result use the clobbered %8
operand for temporary comparison result and set the bool operand at the
end of the definition to true (in this case the immediate value of 1).
Also to ensure that the bool operand is 0 in all other cases the first
instruction which is intended as a clear of the carry bit is reused to
set the bool operand to 0 at the same time as clearing the carry bit.
And finally the jump offsets were updated

Additional to the behaviour change this change defines the side-effects
of the atomic_compare_and_swap. Specifically the side effects where the
bool and val operands are modified/set based on the value of the memory
content. This prevents certain optimization behaviour from incorrectly
optimizing away code. An example of this is the snippet below, where in
certain cases the comparison is optimized away entirely.

  mem = 2;
  if (atomic_compare_and_swap(&mem, ...) == 2)
    ...

This change resolves the following failing testcases for MicroBlaze:
 * gcc.dg/atomic/c11-atomic-exec-3.c
 * gcc.dg/atomic/stdatomic-compare-exchange-3.c
 * gcc.dg/atomic/stdatomic-exchange-1.c
 * gcc.dg/atomic/stdatomic-exchange-2.c
 * gcc.dg/atomic/stdatomic-exchange-3.c
 * gcc.dg/atomic/stdatomic-op-1.c
 * gcc.dg/atomic/stdatomic-op-2.c
 * gcc.dg/atomic/stdatomic-op-3.c
 * gcc.dg/atomic/stdatomic-op-5.c
 * gcc.dg/stdatomic-vm.c

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Upstream-Status: Unsubmitted