Fix the Microblaze crash with msmall-divides flag
Fix the Microblaze crash with msmall-divides flag Compiler is crashing
when we use msmall-divides and mxl-barrel-shift flag. This is because
when use above flags microblaze_expand_divide function will be called
for division operation. In microblaze_expand_divide function we are
using sub_reg but MicroBlaze doesn't have subreg register due to this
compiler was crashing. Changed the logic to avoid sub_reg call

Signed-off-by: Nagaraju Mekala <nagaraju.mekala@xilinx.com>
Signed-off-by: Mahesh Bodapati <mbodapat@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>
Patch-Id: 11
Upstream-Status: Pending