Inline Expansion of fsqrt builtin
Inline Expansion of fsqrt builtin. The changes are made in the patch for
the inline expansion of the fsqrt builtin with fqrt instruction. The
sqrt math function takes double as argument and return double as
argument. The pattern is selected while expanding the unary op through
expand_unop which passes DFmode and the DFmode pattern was not there
returning zero. Thus the sqrt math function is not inlined and expanded.
The pattern with DFmode argument is added. Also the source and
destination argument is not same the DF through two different
consecutive registers with lower 32 bit is the argument passed to sqrt
and the higher 32 bit is zero. If the source and destinations are
different the DFmode 64 bits registers is not set properly giving the
problem in runtime. Such changes are taken care in the implementation of
the pattern for DFmode for inline expansion of the sqrt.

ChangeLog:

2015-06-16  Ajit Agarwal  <ajitkum@xilinx.com>
            Nagaraju Mekala  <nagaraju.mekala@xilinx.com>

	* config/microblaze/microblaze.md (sqrtdf2): New
	pattern.

Signed-off-by: Nagaraju Mekala <nagaraju.mekala@xilinx.com>
Signed-off-by: Ajit Agarwal <ajitkum@xilinx.com>
Signed-off-by: Mahesh Bodapati <mbodapat@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>
Patch-Id: 24
Upstream-Status: Pending