Update heuristics for loop-invariant for address arithmetic.
The changes are made in the patch to update the heuristics for loop
invariant for address arithmetic. The heuristics is changed to calculate
the estimated register pressure cost when ira based register pressure is
not enabled. The estimated register pressure cost modifies the existing
calculation cost associated to perform the Loop invariant code motion
for address arithmetic.

ChangeLog:
2015-06-17  Ajit Agarwal  <ajitkum@xilinx.com>
            Nagaraju Mekala  <nmekala@xilinx.com>

	* loop-invariant.c (gain_for_invariant): update the
	heuristics for estimate_reg_pressure_cost.

Signed-off-by: Nagaraju Mekala <nagaraju.mekala@xilinx.com>
Signed-off-by: Ajit Agarwal <ajitkum@xilinx.com>
Signed-off-by: Mahesh Bodapati <mbodapat@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>
Patch-Id: 25
Upstream-Status: Pending