microblaze.md: Improve 'adddi3' and 'subdi3' insn definitions
Change adddi3 to handle DI immediates as the second operand, this
requires modification to the output template however reduces the need to
specify seperate templates for 16-bit positive/negative immediate
operands. The use of 32-bit immediates for the addi and addic
instructions is handled by the assembler, which will emit the imm
instructions when required. This conveniently handles the optimizable
cases where the immediate constant value does not need the higher half
words of the operands upper/lower words.

Change the constraints of the subdi3 instruction definition such that it
does not match the second operand as an immediate value. This is because
there is no definition to handle this case nor is it possible to
implement purely with instructions as microblaze does not provide an
instruction to perform a forward arithmetic subtraction (it only
provides reverse 'rD = IMM - rA').

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Mahesh Bodapati <mbodapat@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>
Patch-Id: 26
Upstream-Status: Pending