Better register pressure estimate for loop invariant code motion
Calculate the loop liveness used for regs for calculating the register
pressure in the cost estimation.  Loop liveness is based on the
following properties.  We only need to find the set of objects that are
live at the birth or the header of the loop. We don't need to calculate
the live through the loop by considering live in and live out of all the
basic blocks of the loop. This is based on the point that the set of
objects that are live-in at the birth or header of the loop will be
live-in at every node in the loop.

If a v live is out at the header of the loop then the variable is
live-in at every node in the loop. To prove this, consider a loop L with
header h such that the variable v defined at d is live-in at h. Since v
is live at h, d is not part of L. This follows i from the dominance
property, i.e. h is strictly dominated by d. Furthermore, there exists a
path from h to a use of v which does not go through d. For every node p
in the loop, since the loop is strongly connected and node is a
component of the CFG, there exists a path, consisting only of nodes of L
from p to h. Concatenating these two paths proves that v is live-in and
live-out of p.

Calculate the live-out and live-in for the exit edge of the loop. This
patch considers liveness for not only the loop latch but also the
liveness outside the loops.

ChangeLog:

2016-01-22  Ajit Agarwal  <ajitkum@xilinx.com>

	* loop-invariant.c (find_invariants_to_move): Add the logic of
	  regs_used based on liveness.
	* cfgloopanal.c (estimate_reg_pressure_cost): Update the heuristics
	  in presence of call_p.

Signed-off-by: Ajit Agarwal <ajitkum@xilinx.com>
Signed-off-by: Mahesh Bodapati <mbodapat@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>
Patch-Id: 29
Upstream-Status: Pending