Fix internal compiler error with msmall-divides
This patch will fix the internal error microblaze_expand_divide function
which comes because of rtx PLUS where the mem_rtx is of type SI and the
operand is of type QImode. This patch modifies the mem_rtx as QImode and
Plus as QImode to fix the error.

ChangeLog:

2016-02-23  Nagaraju Mekala  <nagaraju.mekala@xilinx.com>
            Ajit Agarwal  <ajitkum@xilinx.com>

	* microblaze.c (microblaze_expand_divide): Update

Signed-off-by: Nagaraju Mekala <nagaraju.mekala@xilinx.com>
Signed-off-by: Ajit Agarwal <ajitkum@xilinx.com>
Signed-off-by: Mahesh Bodapati <mbodapat@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>
Patch-Id: 31
Upstream-Status: Pending