Add new bit-field instructions
This patches adds new bsefi and bsifi instructions. BSEFI- The
instruction shall extract a bit field from a register and place it
right-adjusted in the destination register. The other bits in the
destination register shall be set to zero BSIFI- The instruction shall
insert a right-adjusted bit field from a register at another position in
the destination register. The rest of the bits in the destination
register shall be unchanged

ChangeLog:

2016-02-03  Nagaraju Mekala <nagaraju.mekala@xilinx.com>

	* microblaze.md (Update): Added new patterns

Signed-off-by: Nagaraju Mekala <nagaraju.mekala@xilinx.com>
Signed-off-by: Mahesh Bodapati <mbodapat@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>
Patch-Id: 33
Upstream-Status: Pending