MicroBlaze fix signed bit fields with bit field instructions
The 'extv' definition is expected to sign extended the result based on
the width of the bit field.

  https://gcc.gnu.org/onlinedocs/gccint/Standard-Names.html#index-extvm-instruction-pattern

The MicroBlaze 'bsefi' instruction does not sign extended, it zero
extends. There is no option for the instruction to sign extended the
result and no simple instruction or expression to implement a variant
length sign extend (only sext8/sext16 instructions exist).

As such these definitions needs to be changed to the zero extended
variant of 'extv' which is 'extzv'. This change updates the existing
definitions to allow for signed bit fields to function correctly and be
sign extended.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Upstream-Status: Pending