MicroBlaze improve 'adddi3' handling of immediates
Improve adddi3 to handle DI immediates as the third operand, this
requires modification to the output template. This change however
removes the need to specify separate templates for positive/negative
immediates.

The new definition trades of specific optimization cases for 16-bit
immediates for general coverage of all immediates from 16-bit to DI. The
asssmbler can conveniently optimize out the imm instruction for cases
where the imm instruction is not required (e.g. immediates that are
negative).

Additionally this change resolves the issue where the third operand is
using the 'arith_operand32' predicate which only allows matching of
32-bit immediates.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
---
Upstream-Status: Unsubmitted