WIP: MicroBlaze correct subdi3 predicate (needed?)
Change the predicate of the subdi3 instruction definition such that it
does not use the arith_operand32 predicate which is for
the third
operand as an immediate value. This is because there is no definition to
handle this case (rsub/rsubc do not handle immediate values) and it is
possible to implement purely with instructions as microblaze does not
provide an instruction to perform a forward arithmetic subtraction (it
only provides reverse 'rD = IMM - rA').

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>